refactor(invoice-service): use big decimals instead of doubles (55-8843)

+++ invoice-service/src/main/java/com/team28/booking/invoice/dto/AppliedDiscountDTO.java
@@ -1,5 +1,6 @@
 package com.team28.booking.invoice.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.team28.booking.invoice.model.Discount;
@@ -7,6 +8,6 @@ import com.team28.booking.invoice.model.Discount;
 public record AppliedDiscountDTO(
         String discountCode,
         Discount.DiscountType discountType,
-        Double discountApplied,
+        BigDecimal discountApplied,
         LocalDateTime appliedAt
 ) {}
